created three tiers of radio sending, still requires testing
